(file-logger): add default finder impl
