--- packages/stream_video/lib/src/logger/impl/composite_logger.dart
@@ -1,20 +1,0 @@
-import '../stream_logger.dart';
-
-class CompositeStreamLogger extends StreamLogger {
-  const CompositeStreamLogger(this.children);
-
-  final List<StreamLogger> children;
-
-  @override
-  void log(
-    Priority priority,
-    String tag,
-    MessageBuilder message, [
-    Object? error,
-    StackTrace? stk,
-  ]) {
-    for (final child in children) {
-      child.log(priority, tag, message, error, stk);
-    }
-  }
-}
