--- packages/stream_video/lib/src/logger/impl/silent_logger.dart
@@ -1,16 +1,0 @@
-import '../stream_logger.dart';
-
-class SilentStreamLogger extends StreamLogger {
-  const SilentStreamLogger();
-
-  @override
-  void log(
-    Priority priority,
-    String tag,
-    MessageBuilder message, [
-    Object? error,
-    StackTrace? stk,
-  ]) {
-    /* no-op */
-  }
-}
